integration of insertBodies instead of insertChain
